+++ DataLayer/DBKlasser/Intäktsbudget.cs
@@ -20,15 +20,15 @@ namespace DataLayer
 
         public bool Låst { get; set; }
 
-        public decimal Avtal { get; set; }
+        public double Avtal { get; set; }
 
         public bool GradA { get; set; }
 
-        public decimal Tillägg { get; set; }
+        public double Tillägg { get; set; }
 
         public bool GradT { get; set; }
 
-        public decimal Budget { get; set; }
+        public double Budget { get; set; }
 
         public int Tim { get; set; }
