+++ DataLayer/DTO/KostnadsbudgetProduktDTO.cs
@@ -10,7 +10,7 @@ namespace DataLayer
     {
         public string ProduktID { get; set; }
         public string Produkt { get; set; }
-        public decimal Summering { get; set; } 
+        public double Summering { get; set; } 
         public int Avdelning_AvdelningID { get; set; }
     }
 }
