+++ PresentationLayer1/Forms/frmKostnadsbudgetering.cs
@@ -138,7 +138,7 @@ namespace PresentationLayer1.Forms
             KonstnadsbudgetPersonalDTO personal = (KonstnadsbudgetPersonalDTO)dgvÖvre.CurrentRow.DataBoundItem;
             var pers = personal.PersonalID;
             var produkt = cbProdukt.SelectedItem.ToString();
-            decimal andel = decimal.Parse(tbAndel.Text);
+            double andel = double.Parse(tbAndel.Text);
 
             var pp = businessManager.GetAllPersonal();
 
@@ -150,15 +150,15 @@ namespace PresentationLayer1.Forms
 
             //***********************************Behöver räknas för placeringen********************
 
-            decimal räknaprocent;
-            decimal räkna2;
-            decimal r2;
-            decimal n;
-            decimal nyplaceringsandel;
+            double räknaprocent;
+            double räkna2;
+            double r2;
+            double n;
+            double nyplaceringsandel;
 
 
             räkna2 = andel;
-            räknaprocent = räkna2 / 100m;
+            räknaprocent = räkna2 / (double) 100m;
 
             n = query;
 
@@ -184,7 +184,7 @@ namespace PresentationLayer1.Forms
 
 
 
-            decimal nyandeltest = fördeladandel + nyplaceringsandel;      /* DENNA FÅR INTE VARA MINUS */   
+                double nyandeltest = fördeladandel + nyplaceringsandel;      /* DENNA FÅR INTE VARA MINUS */   
             
 
 
@@ -216,7 +216,7 @@ namespace PresentationLayer1.Forms
         {
             PersonalProduktDTO perspro = (PersonalProduktDTO)dgvNedre.CurrentRow.DataBoundItem;
             var pers = perspro.Personal;
-            decimal andel = perspro.Placeringsandel;
+            double andel = perspro.Placeringsandel;
             var produkt = perspro.Produkt;
 
             businessManager.RemovePlaceringProdukt(pers, produkt, andel);
